started adding support for the chat widget...
(who knew there was chat?  I didn't, until I got a notification
 to "@all" in channel "#general", which notified all users on the
 site and got a bunch of people upset)

I think it's a brand new feature, because the chat function
in the page header doesn't appear in my screenshots from just
a couple days ago.

Anyway, the chat widget is a little awkward, and not really done yet,
but this at least improves it quite a bit from default.